requirements.txt
torch==2.1.1
torchvision==0.16.1
hydra-core==1.3.2
loguru==0.7.2
tqdm==4.66.1
jsonargparse==4.27.0
pillow==10.1.0
numpy==1.26.2